--- automatic rebase to remote; old local SHA: 28c0c25226e448069281c05fefbdda3de3364a9f ---

+++ src/hartuTofas/MessageHandler.java
@@ -44,7 +44,7 @@ public class MessageHandler {
         public int ioState;
         public int tool;
         public int base;
-        public int speedOverride;
+        public float speedOverride;
 
         public boolean programCall;
         public String targetPoints;
